[audio hal] using int64 type to calculate the sleep time
the old codes: time = bytes * 1000000 will lead to overflow if use
int type, this will lead to speed play

+++ tinyalsa_hal/audio_hw.c
@@ -2406,7 +2406,7 @@ final_exit:
     }
     if (ret != 0) {
         ALOGV("AudioData write  error , keep slience! ret = %d", ret);
-        usleep(bytes * 1000000 / audio_stream_out_frame_size(stream) /
+        usleep((int64_t)bytes * 1000000ll / audio_stream_out_frame_size(stream) /
                out_get_sample_rate(&stream->common));
     }
